src/tutorial/urls.py
from django.conf.urls import url, include
from django.contrib import admin
from django.shortcuts import render
from rest_framework import routers
from tutorial.todo import views

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)
router.register(r'todos', views.TodoViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    url(r'index\.html$', lambda request: render(request, 'index.html'), name='index'),
    url(r'^', include(router.urls)),
    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    url(r'^admin/', admin.site.urls),
]
